+++ src/frontend/settings/actions/index.tsx
@@ -1,0 +1,6 @@
+export const loadUserInfo = (user: any) => {
+    return {
+        type: "LOAD_INFO",
+        loadUserInfo: user
+    }
+};